BUG: Wrong data type disallowing access to array, causing segfaults on 64 bit

+++ Code/Numerics/Statistics/itkMersenneTwisterRandomVariateGenerator.h
@@ -484,8 +484,7 @@ inline void
   // Made clearer and faster by Matthew Bellew (matthew.bellew@home.com)
   
   // get rid of VS warning
-  register IntegerType index = static_cast< IntegerType >(
-      M-MersenneTwisterRandomVariateGenerator::StateVectorLength);
+  register int index = M-MersenneTwisterRandomVariateGenerator::StateVectorLength;
     
   register IntegerType *p = state;
   register int i;
